+++ R/ResamplingCV.R
@@ -78,7 +78,11 @@ ResamplingCV = R6Class("ResamplingCV", inherit = Resampling,
     },
 
     deep_clone = function(name, value) {
-      if (name == "instance") copy(value) else value
+      switch(name,
+        "instance" = copy(value),
+        "param_set" = value$clone(deep = TRUE),
+        value
+      )
     }
   )
 )